chore: move interaction/ninode stuff around
