+++ src/Thread/Interaction/NodeLegacy/LegacyNiMath.cpp
