+++ src/Thread/Interaction/NodeLegacy/LegacyNiMath.h
